KAFKA-14589 Move abstract tests to new files
